feat(app.js): add authorization to Front End. Prevent 'log out' when user refreshes page
add authorization headers to every fetch call method (put, post, get) for the app. add session
functionality through the componentDidMount() function. If there is a token in localStorage, then
automatically sign in and fetch the user's data. Since the fetch calls are in the React function
componentDidMount() and it's located in App.js, the function will run as soon as the DOM content is
loaded and will run everytime. Now users won't have to log in if they refresh the page.